feat: Add index.html and ImagesModule
This commit adds a new `index.html` file to the `public` directory, providing a basic HTML structure for the application's landing page.

Additionally, it introduces the `ImagesModule` to the project, which includes the `ImagesService` and `ImagesController` for handling image-related functionality.

The changes also include updating the `package.json` file to add the `@nestjs/serve-static` dependency, which will be used by the `ImagesModule` to serve static image files.

Lastly, two new test files, `images.service.spec.ts` and `images.controller.spec.ts`, have been added to test the functionality of the `ImagesService` and `ImagesController` respectively.